+++ src/index.jsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 
 import { App } from './lib/components/App/App';
-import GsLogoCircle from './lib/images/gs_logo_circle.svg';
+// import GsLogoCircle from './lib/images/gs_logo_circle.svg';
 
 const root = createRoot(document.getElementById('root'));
 
@@ -11,7 +11,8 @@ root.render(
   <React.StrictMode>
     <HashRouter>
       <App
-          appLogo={GsLogoCircle}
+          // appLogo={GsLogoCircle}
+          appLogo={"gs_logo_circle.svg"}
       />
     </HashRouter>
   </React.StrictMode>