+++ src/index.tsx
@@ -5,7 +5,9 @@ import { HashRouter } from 'react-router-dom';
 const reactDomClient = require('react-dom/client');
 
 const app = require('./lib/components/App/App');
-const GsLogoCircle = require('./lib/images/gs_logo_circle.svg');
+
+// This way to import the .svg files doesn't work on prod environents...
+// const GsLogoCircle = require('./lib/images/gs_logo_circle.svg');
 
 const root = reactDomClient.createRoot(document.getElementById('root'));
 
@@ -13,7 +15,8 @@ root.render(
   <React.StrictMode>
     <HashRouter>
       <app.App
-        appLogo={GsLogoCircle}
+          // appLogo={GsLogoCircle}
+          appLogo={"gs_logo_circle.svg"}
       />
     </HashRouter>
   </React.StrictMode>